requirements.txt
Django==3.1
Pillow
psycopg2-binary==2.8.5
djangorestframework==3.11
djangorestframework-simplejwt
twilio
django-phonenumber-field==5.0.0
phonenumbers==8.12.19
python-decouple==3.4